src/bronfood/api/restaurants/views.py
from rest_framework import viewsets

from bronfood.core.restaurants.models import Dish, Menu, Restaurant, Tag

from .serializers import (
    DishSerializer, MenuSerializer, RestaurantSerializer, TagSerializer,
)


class RestaurantViewSet(viewsets.ModelViewSet):
    queryset = Restaurant.objects.all()
    serializer_class = RestaurantSerializer


class MenuViewSet(viewsets.ModelViewSet):
    queryset = Menu.objects.all()
    serializer_class = MenuSerializer


class TagViewSet(viewsets.ModelViewSet):
    queryset = Tag.objects.all()
    serializer_class = TagSerializer


class DishViewSet(viewsets.ModelViewSet):
    queryset = Dish.objects.all()
    serializer_class = DishSerializer
