requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==4.2.10
cryptography==41.0.7
defusedxml==0.8.0rc2
django-cors-headers==4.3.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-yasg==1.21.7
flake8==7.0.0
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.11.2
platformdirs==4.0.0
pluggy==1.4.0
psycopg2-binary==2.9.6
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
PyJWT==2.8.0
pytest==8.0.0
pytest-django==4.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
Pillow==10.1.0
uritemplate==4.1.1
urllib3==2.1.0
